+++ src/components/Feedback/Feedback.test.tsx
@@ -1,7 +1,9 @@
 import React from "react";
-import { render, screen } from "@testing-library/react";
+import { render } from "../../__tests__/testUtils/render";
+import { screen } from "@testing-library/react";
 import Feedback from "./Feedback";
 import userEvent from "@testing-library/user-event";
+jest.mock("next/router", () => require("next-router-mock"));
 
 describe("Feedback", () => {
   beforeEach(() => {